Mms: Notify user of Send SMS failure due to FDN
BZ: 48558

Upon Send SMS failure due to destination number or
service center number not present in the FDN list,
failure is not notified to the MessageNotification.
Due to this Sms remains in Sending state even
after the SMS send has failed.

This patch fixes the issue by notifying the message
failure and also moving the message to respective folder.

Change-Id: Idab7f5816c78e1b334a7033e89b38510003020f0
Orig-Change-Id: Idf960f87d43aa43e9b692833c6df098be53308bc
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/76465
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>