[PORT FROM MAIN]MMS: fix CRASH when the address contains only spaces
BZ: 56634

Rootcause: If the address contains spaces only,when Mms
format this address,it will get an address string the length
of which is -1. So crash happen when call span(object,0,-1,flag)
function.

Solution: When user press back key to restore the message.
This patch will check the address is valite or not, if it is
invalite,it show a dialog to tell user he can't restore the message.
So it avoid crashing when enter the message again.

Feature: AOSP-Messaging
Category: aosp improvement
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I14ce84f1827ec0dbe38098f279451f046e053ba7
Orig-Change-Id: I1eee050b578142e255792401a287c6e47a2d98ab
Signed-off-by: jli119X <jianpingx.li@intel.com>