Mms: fix CRASH when the address contains only spaces
BZ: 56634

Rootcause: If the address contains spaces only,when Mms
format this address,it will get an address string the length
of which is -1. So crash happen when call span(object,0,-1,flag)
function.

Solution: When user press back key to restore the message.
This patch will check the address is valite or not, if it is
invalite,it show a dialog to tell user he can't restore the message.
So it avoid crashing when enter the message again.

Change-Id: I1eee050b578142e255792401a287c6e47a2d98ab
Signed-off-by: jli119X <jianpingx.li@intel.com>
Reviewed-on: http://android.intel.com:8080/76492
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>