Mms: On attachment size exceeded, show error message
BZ: 38953

When forwarding an MMS message which is more than
the maximum allowed size, message size is checked
by a call to checkMessageSize. Upon message size
less than 0 or greater than allowed maximum message
size, ExceedMessageSizeException is thrown. Any
error's occuring while sending MMS, mms app is
implemented to crash.

With this patch, atleast ExceedMessageSizeException
will be catched and an error message will be shown
to the user instead of allowing the mms app to crash.

Change-Id: I7880e80c3676c6570141c164cfd9817a973b44a5
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/76480
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>