[Port From R4.1]MMS: Messaging ANR when share burst pictures
BZ: 71010

The slide list will be created in the onCreate function of
SlideshowEditActivity.java.
It takes about 5 seconds to decode the 10 big images.
The region of the image in the listitem is smaller than the size of
image.(such as 48*48, but the size of image maybe 640*480)
In google's design, it will try to create a bitmap according the size of image.
In new design, we will create a smaller bitmap according to the size of region.
It can save the memory and time.

Change-Id: Ia705c66fd9fde5db3e3f708424010b8562369ac7
Orig-Change-Id: I5da50e5bcec9c925f7d887e53fe467ed6daaf650
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/88119
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>