MMS: Need touch back key twice to return to the MMS edit Activity.
BZ: 75289

Display MMS edit, touch "Edit" button twice.
Need to touch backkey twice if we want to return to the MMS edit.
Disable all of the buttons when click any button except the replace button.
(Because the replace button will raise a dialog in the activity).
And reset to enable all of the buttons when the activity onresume.

Change-Id: I6ddd7ec9efa74c1b2115a3feb6777f006aae1a04
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/84314
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>