[PORT FROM MAIN]MMS: Need touch back key twice to return to the MMS edit Activity.
BZ: 75289

Display MMS edit, touch "Edit" button twice.
Need to touch backkey twice if we want to return to the MMS edit.
Disable all of the buttons when click any button except the replace button.
(Because the replace button will raise a dialog in the activity).
And reset to enable all of the buttons when the activity onresume.

Orig-Change-Id: I6ddd7ec9efa74c1b2115a3feb6777f006aae1a04
Category: aosp improvement
Domain: AOSP-Messaging
Origin: internal
Upstream-Candidate: yes
Change-Id: Ic03dd5f72eae86e2f5c50be02fd60781a824a341
Signed-off-by: Bin Yang <bin.y.yang@intel.com>