[PORT FROM MAIN]MMS: Need touch back key twice to return to the MMS edit Activity.
BZ: 75289

Display MMS edit, touch "Edit" button twice.
Need to touch backkey twice if we want to return to the MMS edit.
Disable all of the buttons when click any button except the replace button.
(Because the replace button will raise a dialog in the activity).
And reset to enable all of the buttons when the activity onresume.

Feature: AOSP-Messaging
Category: aosp improvement
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: I4cc63bdef0710f765a9f80e1b5ef5f770b59a233
Orig-Change-Id: I6ddd7ec9efa74c1b2115a3feb6777f006aae1a04
Signed-off-by: Bin Yang <bin.y.yang@intel.com>