[PORT FROM MAIN]MMS: Fix exception due to closed cursor access
BZ: 77629

The mCursor is closed but not set with null.
The onPrepareOptionsMenu function use the closed cursor resulting in the crash.
This patch fix the issue by seting the mCursor with null after it is closed
in refreshMessageList function.

Feature: AOSP-Messaging
Category: aosp improvement
Severity: medium
Origin: internal
Upstream-Candidate: yes

Change-Id: If597a1c4f423f50c4459a62a67e011b53f0989f2
Orig-Change-Id: I3ff079d358af2f3d41ae60a283c3b679c7eeb556
Signed-off-by: Bin Yang <bin.y.yang@intel.com>