MMS: The contact does not exist appears when click the contact
BZ: 80277

A inforamtion about "The contact does not exist" will be prompted
when user clicks the contact icon of the receiptor in one thread.
The root cause is that there is an error logic to check the contact is self or not.
Correct the logic to display the contact information of receiptor.

Change-Id: I88ea7047bbe091a8883ba6a53ac11959b64c70ff
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/86682
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>