MMS: Display of slide is paused.
BZ: 81319

The root cause is the user push the play button twice very quickly.
Every play request will show a SlideShowActivity.
The smilplayer is paused when the second SlideshowActivity is showing.
Fix this issue by disabling the play button when user push the play button.
So the second request of showing the SlideshowActivity can't be made.

Change-Id: I81bbe1d9b6598bdf8cc58816043966c41dde42b5
Orig-Change-Id: Ieecf913fafe2b44598cdcaa101974fed5631bd91
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/90384
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>