[PORT FROM R4.1] MMS: The message details performs wrong type
BZ: 83120

The root cause is that the wrong cursor is returned
when user want to check the detailed information of a mms.
The messages are divided into two classes: mms and sms.
One mms and one sms maybe have the same id but have different message type.
Add a type check in getCursorForItem function, so we can get the
right cursor.

Change-Id: Ie6f6615b08f203ddb3df907b05e73622729d6da5
Orig-Change-Id: I5f5268181686fccfd9a34c2993d2da56ad796f74
Signed-off-by: Bin Yang <bin.y.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/92674
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>