[PORT FROM MAIN] MMS: The message details performs wrong type
BZ: 83120

The root cause is that the wrong cursor is returned
when user want to check the detailed information of a mms.
The messages are divided into two classes: mms and sms.
One mms and one sms maybe have the same id but have different message type.
Add a type check in getCursorForItem function, so we can get the
right cursor.

Orig-Change-Id: Ie6f6615b08f203ddb3df907b05e73622729d6da5
Category: aosp improvement
Domain: AOSP-Messaging
Origin: internal
Upstream-Candidate: yes
Change-Id: I48cbc9e437b1c33788efceff49acc679151dcb63
Signed-off-by: Bin Yang <bin.y.yang@intel.com>