Mms: Not all MMS is sent or received.
BZ: 82238

When there are multiple sending and receiving transactions, there
is possibility if a case in which have two transaction in processing
list and only one will be processed.

This patch adds the transaction to the Pending list when there is
already a transaction in processing state and also doesn't disconnect
mms connectivity if there are transactions in processing state.

Change-Id: I8d2505418c016f7117113e442aa6adf4fa049f9e
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/97901
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>