Mms: Loading message and draft immediately when opening message list
BZ: 79275

To speed up message and draft loading process. There are two cases here,
one is there is draft, in this case, it will introduce jank which message
list shows then the IME window shown, message list will be resize and
move up; Other is that no draft, in this case, IME doesn't appear,
so no jank. The former case is minor case, so we loading draft and
message immediately to optimize the UX.

Change-Id: I3e79578549050e9fd4a9eca4f478385975247e3c
Signed-off-by: gche50X <gangx.a.chen@intel.com>
Reviewed-on: http://android.intel.com:8080/118103
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>