[PORT FROM MAIN] Mms: Not all MMS is sent or received
BZ: 82238
When there are multiple sending and receiving transactions,
there is possibility if a case in which there are two
transactions in processinglist and only one will be processed.

This patch adds the transaction to the pending list when there
is already a transaction in processing state and also doesn't
disconnect mms connectivity if there are transactions in
processing state.

Category: AOSP Improvement
Domain: Telephony-Mms
Origin: Internal
Upstream-Candidate: Yes

Change-Id: If247e7e6624dda3bde9dfeaa5c8465762b2769c8
Orig-Change-Id: I8d2505418c016f7117113e442aa6adf4fa049f9e
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/125030
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>