[PORT FROM MAIN] Mms: Not all MMS is sent or received.
BZ: 82238

When there are multiple sending and receiving transactions,
there is possibility if a case in which there are two
transactions in processinglist and only one will be processed.

This patch adds the transaction to the pending list when there
is already a transaction in processing state and also doesn't
disconnect mms connectivity if there are transactions in
processing state.

Category: aosp improvement
Domain: AOSP-Messaging
Origin: internal
Upstream-Candidate: yes

Orig-Change-Id: If247e7e6624dda3bde9dfeaa5c8465762b2769c8
Change-Id: I77c60726dbdb43280057bd3db60d9649b15a17f2
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>