[PATCH] Mms: Preferred sending SMS messages SIM
When sending a sms, promise a identifier to judge if it is in
ask every time mode. It can be set in Settings, and if it is in
ask every time mode, Messaging will pop a dialog to make the user
choose a sim to send sms. Telephony will send the sms using the subid
on the basis of the user choose.

Change-Id: If8badb954f66d7c52d14a78d78bc5a40e9e3418c
Category: AOSP improvement
Domain:  Mms
Origin: Internal
Upstream-Candidate: no
Signed-off-by: jialinx.dong <jialinx.dong@intel.com>